requirements.txt
apimatic-core==0.2.3
apimatic-core-interfaces==0.1.3
apimatic-requests-client-adapter==0.1.4
asgiref==3.7.2
CacheControl==0.12.11
certifi==2023.5.7
charset-normalizer==3.1.0
deprecation==2.1.0
Django==4.2.1
django-contrib-postgres==0.0.1
djangorestframework==3.14.0
enum34==1.1.10
gis==0.2.1
idna==3.4
jsonpickle==3.0.1
jsonpointer==2.3
msgpack==1.0.5
packaging==23.1
psycopg2-binary==2.9.6
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
squareup==27.0.0.20230517
urllib3==2.0.2
uuid==1.30
